+++ solve.h
@@ -12,8 +12,8 @@ double Norm (double *a, int v, int h);
 void Multi (double *x, double *y, double *res, int vx, int hx, int hy);
 void Plus (double *x, double *y, double *res, int v, int h);
 void Minus (double *x, double *y, double *res, int v, int h);
-int Inverse (double *matrix, double *inversed, int matrix_n, int matrix_norm, int *ind);
-int Solve (double *matrix, double *inversed_matrix, int matrix_n, int block_m, int matrix_norm, int *ind);
+int Inverse (double *matrix, double *inversed, int matrix_n, double matrix_norm, int *ind);
+int Solve (double *matrix, double *inversed_matrix, int matrix_n, int block_m, double matrix_norm, int *ind);
 void ReplaceLines (double *matrix, int matrix_n, int block_m, int *ind);
 
 #endif
